requirements.txt
# ============================================
# CareerCraft AI - Dependencies
# ============================================

# --- Core LLM & Agent Framework ---
langchain>=0.3.14
langchain-core>=0.3.28
langchain-community>=0.3.14
langchain-groq>=0.2.4
langgraph>=0.2.60
langchain-huggingface>=0.0.3
langchain-tavily>=0.0.1

# --- Embeddings ---
sentence-transformers>=3.3.1

# --- Vector Database (RAG) ---
chromadb>=0.5.23
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0

# --- PDF & Document Parsing ---
pdfplumber>=0.11.4
python-docx>=1.1.2

# --- Web Search Tool ---
tavily-python>=0.5.0

# --- UI Framework ---
chainlit>=1.3.2
streamlit>=1.31.0

# --- Persistence ---
sqlalchemy>=2.0.36

# --- Validation ---
pydantic>=2.10.4
pydantic-settings>=2.7.1

# --- Charts & Visualization ---
plotly>=5.24.1

# --- Utilities ---
python-dotenv>=1.0.1

# --- Testing ---
pytest>=8.3.4
pytest-asyncio>=0.24.0
